+++ actionpack/lib/action_controller/metal/rendering.rb
@@ -4,7 +4,7 @@ module ActionController
   module Rendering
     extend ActiveSupport::Concern
 
-    RENDER_FORMATS_IN_PRIORITY = [:body, :text, :plain, :html]
+    RENDER_FORMATS_IN_PRIORITY = [:body, :plain, :html]
 
     module ClassMethods
       # Documentation at ActionController::Renderer#render
@@ -82,17 +82,6 @@ module ActionController
       # Normalize both text and status options.
       def _normalize_options(options) #:nodoc:
         _normalize_text(options)
-
-        if options[:text]
-          ActiveSupport::Deprecation.warn <<-WARNING.squish
-          `render :text` is deprecated because it does not actually render a
-          `text/plain` response. Switch to `render plain: 'plain text'` to
-          render as `text/plain`, `render html: '<strong>HTML</strong>'` to
-          render as `text/html`, or `render body: 'raw'` to match the deprecated
-          behavior and render with the default Content-Type, which is
-          `text/html`.
-        WARNING
-        end
 
         if options[:html]
           options[:html] = ERB::Util.html_escape(options[:html])