+++ actionview/lib/action_view/layouts.rb
@@ -426,7 +426,7 @@ module ActionView
     end
 
     def _include_layout?(options)
-      (options.keys & [:body, :text, :plain, :html, :inline, :partial]).empty? || options.key?(:layout)
+      (options.keys & [:body, :plain, :html, :inline, :partial]).empty? || options.key?(:layout)
     end
   end
 end