+++ railties/test/application/middleware_test.rb
@@ -227,9 +227,9 @@ module ApplicationTests
       class ::OmgController < ActionController::Base
         def index
           if params[:nothing]
-            render text: ""
+            render plain: ""
           else
-            render text: "OMG"
+            render plain: "OMG"
           end
         end
       end
@@ -239,7 +239,7 @@ module ApplicationTests
       get "/"
       assert_equal 200, last_response.status
       assert_equal "OMG", last_response.body
-      assert_equal "text/html; charset=utf-8", last_response.headers["Content-Type"]
+      assert_equal "text/plain; charset=utf-8", last_response.headers["Content-Type"]
       assert_equal "max-age=0, private, must-revalidate", last_response.headers["Cache-Control"]
       assert_equal etag, last_response.headers["Etag"]
 
@@ -253,7 +253,7 @@ module ApplicationTests
       get "/?nothing=true"
       assert_equal 200, last_response.status
       assert_equal "", last_response.body
-      assert_equal "text/html; charset=utf-8", last_response.headers["Content-Type"]
+      assert_equal "text/plain; charset=utf-8", last_response.headers["Content-Type"]
       assert_equal "no-cache", last_response.headers["Cache-Control"]
       assert_equal nil, last_response.headers["Etag"]
     end