+++ railties/test/railties/engine_test.rb
@@ -327,7 +327,7 @@ module RailtiesTest
       controller "foo", <<-RUBY
         class FooController < ActionController::Base
           def index
-            render :text => "foo"
+            render plain: "foo"
           end
         end
       RUBY
@@ -341,7 +341,7 @@ module RailtiesTest
       @plugin.write "app/controllers/bar_controller.rb", <<-RUBY
         class BarController < ActionController::Base
           def index
-            render :text => "bar"
+            render plain: "bar"
           end
         end
       RUBY
@@ -436,7 +436,7 @@ YAML
       @plugin.write "app/controllers/admin/foo/bar_controller.rb", <<-RUBY
         class Admin::Foo::BarController < ApplicationController
           def index
-            render text: "Rendered from namespace"
+            render plain: "Rendered from namespace"
           end
         end
       RUBY
@@ -536,7 +536,7 @@ YAML
       controller "foo", <<-RUBY
         class FooController < ActionController::Base
           def index
-            render text: params[:username]
+            render plain: params[:username]
           end
         end
       RUBY
@@ -700,7 +700,7 @@ YAML
           end
 
           def show
-            render text: foo_path
+            render plain: foo_path
           end
 
           def from_app
@@ -712,7 +712,7 @@ YAML
           end
 
           def polymorphic_path_without_namespace
-            render text: polymorphic_path(Post.new)
+            render plain: polymorphic_path(Post.new)
           end
         end
       RUBY
@@ -835,7 +835,7 @@ YAML
       @plugin.write "app/controllers/bukkits/awesome/foo_controller.rb", <<-RUBY
         class Bukkits::Awesome::FooController < ActionController::Base
           def index
-            render :text => "ok"
+            render plain: "ok"
           end
         end
       RUBY
@@ -1220,7 +1220,7 @@ YAML
                 fullpath: \#{request.fullpath}
                 path: \#{request.path}
               TEXT
-              render text: text
+              render plain: text
             end
           end
         end
@@ -1257,7 +1257,7 @@ YAML
       app_file "app/controllers/bar_controller.rb", <<-RUBY
         class BarController < ApplicationController
           def index
-            render text: bukkits.bukkit_path
+            render plain: bukkits.bukkit_path
           end
         end
       RUBY
@@ -1278,7 +1278,7 @@ YAML
       @plugin.write "app/controllers/bukkits/bukkit_controller.rb", <<-RUBY
         class Bukkits::BukkitController < ActionController::Base
           def index
-            render text: main_app.bar_path
+            render plain: main_app.bar_path
           end
         end
       RUBY
@@ -1306,7 +1306,7 @@ YAML
       app_file "app/controllers/bar_controller.rb", <<-RUBY
         class BarController < ApplicationController
           def index
-            render text: bukkits.bukkit_path
+            render plain: bukkits.bukkit_path
           end
         end
       RUBY
@@ -1327,7 +1327,7 @@ YAML
       @plugin.write "app/controllers/bukkits/bukkit_controller.rb", <<-RUBY
         class Bukkits::BukkitController < ActionController::Base
           def index
-            render text: main_app.bar_path
+            render plain: main_app.bar_path
           end
         end
       RUBY